fixing the shape footprint problem and the rotation

+++ src/firefly/java/edu/caltech/ipac/firefly/visualize/WebPlot.java
@@ -1022,8 +1022,8 @@ public class WebPlot {
     @JsNoExport
     public ScreenPt getScreenCoords(ImageWorkSpacePt ipt, float altZLevel) {
         if (ipt==null) return null;
-        return new ScreenPt((int)((ipt.getX())*altZLevel),
-                            (int)((getImageHeight() - ipt.getY()) *altZLevel));
+        return new ScreenPt(((ipt.getX())*altZLevel),
+                            ((getImageHeight() - ipt.getY()) *altZLevel));
     }
 
 //========================================================================================
@@ -1315,7 +1315,7 @@ public class WebPlot {
         CoordinateSys csys= wp.getCoordSys();
         Pt retPt= wp;
         if (csys.equals(CoordinateSys.SCREEN_PIXEL)) {
-            retPt= new ScreenPt((int)wp.getX(), (int)wp.getY());
+            retPt= new ScreenPt(wp.getX(), wp.getY());
         }
         else if (csys.equals(CoordinateSys.PIXEL)) {
             retPt= new ImagePt(wp.getX(), wp.getY());
